requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
scipy>=1.10.0
umap-learn>=0.5.0
hdbscan>=0.8.33
matplotlib>=3.7.0
seaborn>=0.13.0
joblib>=1.3.0
fastapi>=0.110.0
uvicorn>=0.23.0
python-jose[cryptography]>=3.3.0
google-search-results>=2.4.2
python-dotenv>=1.0.0
python-multipart>=0.0.9
bcrypt>=4.0.1
pymupdf>=1.23.0
sqlalchemy>=2.0.0
requests>=2.31.0
azure-storage-blob>=12.19.0
sendgrid>=6.10.0  # Optional: for SendGrid provider
boto3>=1.28.0  # Optional: for AWS SES provider
